+++ src/main/java/snw/rfm/RunForMoney.java
@@ -1,10 +1,10 @@
 /**
  * This file is part of RunForMoney.
- *
+ * <p>
  * RunForMoney is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
- *
+ * <p>
  * RunForMoney is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
- *
+ * <p>
  * You should have received a copy of the GNU General Public License along with RunForMoney. If not, see <https://www.gnu.org/licenses/>.
  */
 
@@ -26,11 +26,13 @@ import snw.rfm.commands.hunter.DeactivateHunterCommand;
 import snw.rfm.commands.team.HunterCommand;
 import snw.rfm.commands.team.LeaveTeamCommand;
 import snw.rfm.commands.team.RunnerCommand;
-import snw.rfm.processor.EventProcessor;
 import snw.rfm.config.GameConfiguration;
-import snw.rfm.game.GameProcess;
 import snw.rfm.config.Preset;
-import snw.rfm.processor.HunterPauseCardItemProcessor;
+import snw.rfm.game.GameProcess;
+import snw.rfm.processor.EventProcessor;
+import snw.rfm.processor.ExitingPickaxeProcessor;
+import snw.rfm.processor.HunterPauseCardProcessor;
+import snw.rfm.tasks.Updater;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -57,54 +59,59 @@ public final class RunForMoney extends JavaPlugin {
         if (Bukkit.getWorld("world") == null) {
             l.sendMessage("[RunForMoney] " + ChatColor.RED + "错误: 世界 'world' 不存在，插件无法加载。请把将用于游戏的地图改名为 'world' 再重试。");
             pmgr.disablePlugin(this);
-        } else {
-            l.sendMessage("[RunForMoney] " + ChatColor.GREEN + "============ Run FOR Money ============");
-            l.sendMessage("[RunForMoney] " + ChatColor.GREEN + "本插件由 SNWCreations @ MCBBS.NET 制作");
-
-            Logger ll = getLogger();
-            ll.info("加载数据...");
-            GameConfiguration.check(); // 2022/2/7 v1.1.5 GameConfiguration 不应该是需要实例化的。
-            Preset.init();
-
-            RFMItems.init();
-
-            ll.info("注册命令...");
-            // region 注册命令
-            // v1.1.0 把本类的 registerCommand 移至 Util 类。
-            Util.registerCommand("start", new StartCommand());
-            Util.registerCommand("forcestop", new ForceStopCommand());
-            Util.registerCommand("hunter", new HunterCommand());
-            Util.registerCommand("runner", new RunnerCommand());
-            Util.registerCommand("leaveteam", new LeaveTeamCommand());
-            Util.registerCommand("endroom", new EndRoomCommand());
-            Util.registerCommand("newgroup", new NewGroupCommand());
-            Util.registerCommand("removegroup", new RemoveGroupCommand());
-            Util.registerCommand("activategroup", new ActivateGroupCommand());
-            Util.registerCommand("deactivategroup", new DeactivateGroupCommand());
-            Util.registerCommand("joingroup", new JoinGroupCommand());
-            Util.registerCommand("leavegroup", new LeaveGroupCommand());
-            Util.registerCommand("activatehunter", new ActivateHunterCommand());
-            Util.registerCommand("deactivatehunter", new DeactivateHunterCommand());
-            Util.registerCommand("resume", new ResumeCommand());
-            Util.registerCommand("grouplist", new GroupListCommand());
-            Util.registerCommand("teamlist", new TeamListCommand());
-            Util.registerCommand("coinlist", new CoinListCommand());
-            Util.registerCommand("rfmitem", new RFMItemCommand());
-            Util.registerCommand("exportcoinlist", new ExportListCommand());
-            // endregion
-
-            // region 注册调试命令
-            // 警告: 以下注册的命令不应该被最终用户使用。
-            Util.registerCommand("forcestart", new ForceStartCommand());
-            // endregion
-
-            ll.info("注册事件处理器...");
-            pmgr.registerEvents(new EventProcessor(), this);
-            pmgr.registerEvents(new HunterPauseCardItemProcessor(), this); // 2022/1/30 笑死，没注册暂停卡的处理器用个鬼哦
-            getLogger().info("加载完成。");
+            return; // 2022/2/9 因为看着 else 不爽，所以我把它删了。
         }
 
+        l.sendMessage("[RunForMoney] " + ChatColor.GREEN + "============ Run FOR Money ============");
+        l.sendMessage("[RunForMoney] " + ChatColor.GREEN + "本插件由 SNWCreations @ MCBBS.NET 制作");
 
+        Logger ll = getLogger();
+        ll.info("加载数据...");
+        GameConfiguration.check(); // 2022/2/7 v1.1.5 GameConfiguration 不应该是需要实例化的。
+        Preset.init();
+
+        RFMItems.init();
+
+        ll.info("注册命令...");
+        // region 注册命令
+        // v1.1.0 把本类的 registerCommand 移至 Util 类。
+        Util.registerCommand("start", new StartCommand());
+        Util.registerCommand("forcestop", new ForceStopCommand());
+        Util.registerCommand("hunter", new HunterCommand());
+        Util.registerCommand("runner", new RunnerCommand());
+        Util.registerCommand("leaveteam", new LeaveTeamCommand());
+        Util.registerCommand("endroom", new EndRoomCommand());
+        Util.registerCommand("newgroup", new NewGroupCommand());
+        Util.registerCommand("removegroup", new RemoveGroupCommand());
+        Util.registerCommand("activategroup", new ActivateGroupCommand());
+        Util.registerCommand("deactivategroup", new DeactivateGroupCommand());
+        Util.registerCommand("joingroup", new JoinGroupCommand());
+        Util.registerCommand("leavegroup", new LeaveGroupCommand());
+        Util.registerCommand("activatehunter", new ActivateHunterCommand());
+        Util.registerCommand("deactivatehunter", new DeactivateHunterCommand());
+        Util.registerCommand("resume", new ResumeCommand());
+        Util.registerCommand("grouplist", new GroupListCommand());
+        Util.registerCommand("teamlist", new TeamListCommand());
+        Util.registerCommand("coinlist", new CoinListCommand());
+        Util.registerCommand("rfmitem", new RFMItemCommand());
+        Util.registerCommand("exportcoinlist", new ExportListCommand());
+        // endregion
+
+        // region 注册调试命令
+        // 警告: 以下注册的命令不应该被最终用户使用。
+        Util.registerCommand("forcestart", new ForceStartCommand());
+        // endregion
+
+        ll.info("注册事件处理器...");
+        pmgr.registerEvents(new EventProcessor(), this);
+        pmgr.registerEvents(new ExitingPickaxeProcessor(), this);
+        ItemRegistry.register(RFMItems.HUNTER_PAUSE_CARD, new HunterPauseCardProcessor());
+
+        getLogger().info("加载完成。");
+
+        if (getConfig().getBoolean("check_update", false)) { // 检查更新
+            new Updater().runTaskAsynchronously(this);
+        }
     }
 
     @Override