+++ src/main/java/snw/rfm/api/events/BaseEvent.java
@@ -1,4 +1,4 @@
-/*
+/**
  * This file is part of RunForMoney.
  *
  * RunForMoney is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
@@ -8,8 +8,25 @@
  * You should have received a copy of the GNU General Public License along with RunForMoney. If not, see <https://www.gnu.org/licenses/>.
  */
 
-/**
- * 此包是整个插件所有事件类的存放地。
- */
+package snw.rfm.api.events;
 
-package snw.rfm.events;
+import org.bukkit.event.Event;
+import org.bukkit.event.HandlerList;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * 本插件所有事件的基类。您不应该监听这个类。
+ */
+public class BaseEvent extends Event {
+    private static final HandlerList handlers = new HandlerList();
+
+    @NotNull
+    @Override
+    public HandlerList getHandlers() {
+        return handlers;
+    }
+
+    public static HandlerList getHandlerList() {
+        return handlers;
+    }
+}