+++ src/main/java/snw/rfm/api/events/GameStopEvent.java
@@ -1,0 +1,34 @@
+/**
+ * This file is part of RunForMoney.
+ *
+ * RunForMoney is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ *
+ * RunForMoney is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with RunForMoney. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package snw.rfm.api.events;
+
+import org.bukkit.entity.Player;
+import org.jetbrains.annotations.Nullable;
+import snw.rfm.game.TeamHolder;
+
+import java.util.Set;
+
+/**
+ * 游戏停止事件。
+ */
+public final class GameStopEvent extends BaseEvent {
+    /**
+     * 获取获胜者集合。
+     * <p>
+     * 其本质是 TeamHolder.getInstance().getRunners() 。
+     *
+     * @return 获胜者集合。
+     */
+    @Nullable
+    public Set<Player> getWinner() {
+        return TeamHolder.getInstance().getRunners();
+    }
+}